requirements.txt
flask
pandas
numpy
matplotlib
scikit-learn
python-dotenv
requests
groq
google-generativeai
werkzeug
openpyxl
gunicorn
flask-cors
flask-limiter